+++ lib/query/query.dart
@@ -8,3 +8,4 @@ export 'zen_query.dart';
 export 'zen_query_cache.dart';
 export 'zen_query_builder.dart';
 export 'zen_query_config.dart';
+export 'zen_mutation.dart';